SGL {SGL}	R Documentation

Fit a GLM with a combination of lasso and group lasso regularization

Description

Fit a regularized generalized linear model via penalized maximum likelihood. The model is fit for a path of values of the penalty parameter. Fits linear, logistic and Cox models.

Usage

SGL(data, index, type = "linear", maxit = 1000, thresh = 0.001, min.frac = 0.1, nlam = 20, gamma = 0.8, standardize = TRUE, verbose = FALSE, step = 1, reset = 10, alpha = 0.95, lambdas = NULL)

Arguments

data
For type="linear" should be a list with $x$ an input matrix of dimension n-obs by p-vars, and $y$ a length $n$ response vector. For type="logit" should be a list with $x$, an input matrix, as before, and $y$ a length $n$ binary response vector. For type="cox" should be a list with x as before, time, an n-vector corresponding to failure/censor times, and status, an n-vector indicating failure (1) or censoring (0).

index
A p-vector indicating group membership of each covariate

type
model type: one of ("linear","logit", "cox")

maxit
Maximum number of iterations to convergence

thresh
Convergence threshold for change in beta

min.frac
The minimum value of the penalty parameter, as a fraction of the maximum value

nlam
Number of lambda to use in the regularization path

gamma
Fitting parameter used for tuning backtracking (between 0 and 1)

standardize
Logical flag for variable standardization prior to fitting the model.

verbose
Logical flag for whether or not step number will be output

step
Fitting parameter used for inital backtracking step size (between 0 and 1)

reset
Fitting parameter used for taking advantage of local strong convexity in nesterov momentum (number of iterations before momentum term is reset)

alpha
The mixing parameter. alpha = 1 is the lasso penalty. alpha = 0 is the group lasso penalty.

lambdas
A user specified sequence of lambda values for fitting. We recommend leaving this NULL and letting SGL self-select values

Details

The sequence of models along the regularization path is fit by accelerated generalized gradient descent.

Value

An object with S3 class "SGL"

beta
A p by nlam matrix, giving the penalized MLEs for the nlam different models, where the index corresponds to the penalty parameter lambda

lambdas
The actual sequence of lambda values used (penalty parameter)

type
Response type (linear/logic/cox)

intercept
For some model types, an intercept is fit

X.transform
A list used in predict which gives the empirical mean and variance of the x matrix used to build the model

lambdas
A user specified sequence of lambda values for fitting. We recommend leaving this NULL and letting SGL self-select values

Author(s)

Noah Simon, Jerry Friedman, Trevor Hastie, and Rob Tibshirani
Maintainer: Noah Simon nsimon@stanford.edu

References

Simon, N., Friedman, J., Hastie, T., and Tibshirani, R. (2011) A Sparse-Group Lasso,
http://www-stat.stanford.edu/~nsimon/SGL.pdf

See Also

cv.SGL

Examples

n = 50; p = 100; size.groups = 10
index <- ceiling(1:p / size.groups)
X = matrix(rnorm(n * p), ncol = p, nrow = n)
beta = (-2:2)
y = X[,1:5] %*% beta + 0.1*rnorm(n)
data = list(x = X, y = y)
fit = SGL(data, index, type = "linear")

[Package SGL version 1.1 Index]