SGL-package {SGL}	R Documentation

Fit a GLM (or cox model) with a combination of lasso and group lasso regularization

Description

Fit a regularized generalized linear model via penalized maximum likelihood. The model is fit for a path of values of the penalty parameter. Fits linear, logistic and Cox models.

Details

Package: 	 SGL

Type: 	 Package

Version: 	 1.0

Date: 	 2012-3-12

License: 	 GPL

LazyLoad: 	 yes

Only 4 functions: SGL cvSGL predictSGL plot.cvSGL

Author(s)

Noah Simon, Jerome Friedman, Trevor Hastie, and Rob Tibshirani

Maintainer: Noah Simon <nsimon@stanford.edu>

References

Simon, N., Friedman, J., Hastie T., and Tibshirani, R. (2011) A Sparse-Group Lasso,
http://www-stat.stanford.edu/~nsimon/SGL.pdf

[Package SGL version 1.1 Index]