plot.cv.SGL {SGL}	R Documentation

plots the cross-validated error curve produced by cv.SGL

Description

Plots the cross-validated error curve, and confidence bounds for each lambda in our regularization path.

Usage

## S3 method for class 'cv.SGL'
plot(x, ...)

Arguments

x
fitted "cv.SGL" object

...
additional arguments to be passed to plot

Details

A cross validated deviance plot is produced. More regularized models are to the right (less regularized to the left)

Author(s)

Noah Simon, Jerome Friedman, Trevor Hastie, and Rob Tibshirani
Maintainer: Noah Simon <nsimon@stanford.edu>

References

Simon, N., Friedman, J., Hastie T., and Tibshirani, R. (2011) A Sparse-Group Lasso,
http://www-stat.stanford.edu/~nsimon/SGL.pdf

See Also

SGL and cv.SGL.

Examples

n = 50; p = 100; size.groups = 10
index <- ceiling(1:p / size.groups)
X = matrix(rnorm(n * p), ncol = p, nrow = n)
beta = (-2:2)
y = X[,1:5] %*% beta + 0.1*rnorm(n)
data = list(x = X, y = y)
cvFit = cvSGL(data, index, type = "linear")
plot(cvFit)

[Package SGL version 1.1 Index]