predictSGL {SGL}	R Documentation

Outputs predicted responses from an SGL model for new observations

Description

Outputs predicted response values for new user input observations at a specified lambda value

Usage

predictSGL(x, newX, lam)

Arguments

x
fitted "SGL" object

newX
covariate matrix for new observations whose responses we wish to predict

lam
the index of the lambda value for the model with which we desire to predict

Details

Predicted outcomes are given

Author(s)

Noah Simon, Jerome Friedman, Trevor Hastie, and Rob Tibshirani
Maintainer: Noah Simon <nsimon@stanford.edu>

References

Simon, N., Friedman, J., Hastie T., and Tibshirani, R. (2011) A Sparse-Group Lasso,
http://www-stat.stanford.edu/~nsimon/SGL.pdf

See Also

SGL and cvSGL.

Examples

n = 50; p = 100; size.groups = 10
index <- ceiling(1:p / size.groups)
X = matrix(rnorm(n * p), ncol = p, nrow = n)
beta = (-2:2)
y = X[,1:5] %*% beta + 0.1*rnorm(n)
data = list(x = X, y = y)
Fit = SGL(data, index, type = "linear")
X.new = matrix(rnorm(n * p), ncol = p, nrow = n)
predictSGL(Fit, X.new, 5)

[Package SGL version 1.1 Index]